modification schema sauce ajout multer

+++ controllers/sauces.js
@@ -1,27 +1,83 @@
 const Sauce = require('../models/sauces');
+const fs = require('fs');
+
+
+
+
 
 //route post
 exports.createSauce = (req, res, next) => {
-    delete req.body._id;
-    const sauce = new sauce({
-        ...req.body
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
+    const sauce = new Sauce({
+        ...sauceObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
+
     sauce.save()
-        .then(() => res.status(201).json({ message: 'sauce enregistré !' }))
-        .catch(error => res.status(400).json({ error }));
-}
+        .then(() => { res.status(201).json({ message: 'Sauce enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
+};
+
+
+
 //route update 
+// exports.modifySauce = (req, res, next) => {
+//     Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+//         .then(() => res.status(200).json({ message: 'sauce modifié !' }))
+//         .catch(error => res.status(400).json({ error }));
+// };
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'sauce modifié !' }))
-        .catch(error => res.status(400).json({ error }));
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete sauceObject._userId;
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Sauce modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
+
 //route delete
+// exports.deleteSauce = (req, res, next) => {
+//     Sauce.deleteOne({ _id: req.params.id })
+//         .then(() => res.status(200).json({ message: 'sauce supprimé !' }))
+//         .catch(error => res.status(400).json({ error }));
+// };
+
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'sauce supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Sauce supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
+
+
 // route getOne
 exports.getOneSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
@@ -31,9 +87,16 @@ exports.getOneSauce = (req, res, next) => {
 
 // route getAll
 
-exports.getAllsauce = (req, res, next) => {
+exports.getAllSauce = (req, res, next) => {
     Sauce.find()
         .then(sauces => res.status(200).json(sauces))
         .catch(error => res.status(404).json({ error }));
 
 };
+
+
+//route post
+
+exports.rateSauce = (req, res, next) => {
+
+}
